Trigger prompt validation
