+++ frontend/src/pages/ActivityReport/Pages/components/__tests__/goalValidator.js
@@ -9,6 +9,7 @@ import {
   OBJECTIVE_TITLE,
   OBJECTIVE_TTA,
   OBJECTIVE_RESOURCES,
+  validatePrompts,
 } from '../goalValidator';
 import {
   GOAL_NAME_ERROR,
@@ -181,6 +182,36 @@ describe('validateGoals', () => {
         const result = unfinishedGoals(goals);
         expect(result).toEqual(false);
       });
+    });
+  });
+
+  describe('validatePrompts', () => {
+    it('returns true if no prompts', async () => {
+      const trigger = jest.fn(() => true);
+      const prompts = [];
+      const result = await validatePrompts(prompts, trigger);
+      expect(result).toBeTruthy();
+    });
+
+    it('returns true if prompts are undefined', async () => {
+      const trigger = jest.fn(() => true);
+      const prompts = undefined;
+      const result = await validatePrompts(prompts, trigger);
+      expect(result).toBeTruthy();
+    });
+
+    it('returns the result of trigger when true', async () => {
+      const trigger = jest.fn(() => true);
+      const prompts = [{ trigger: 'trigger', prompt: 'prompt' }];
+      const result = await validatePrompts(prompts, trigger);
+      expect(result).toBeTruthy();
+    });
+
+    it('returns the result of trigger when false', async () => {
+      const trigger = jest.fn(() => false);
+      const prompts = [{ trigger: 'trigger', prompt: 'prompt' }];
+      const result = await validatePrompts(prompts, trigger);
+      expect(result).toBeFalsy();
     });
   });
